% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{encoder_one_hot}
\alias{encoder_one_hot}
\title{One-Hot Encode Categorical Variables}
\usage{
encoder_one_hot(data)
}
\arguments{
\item{data}{A dataframe containing the dataset to process.}
}
\value{
A dataframe where categorical variables are replaced by one-hot encoded columns.
}
\description{
One-Hot Encode Categorical Variables
}
