% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normaliser_donnees}
\alias{normaliser_donnees}
\title{Normalize/Standardize Numerical Variables}
\usage{
normaliser_donnees(data)
}
\arguments{
\item{data}{A dataframe containing the dataset to process.}
}
\value{
A dataframe with normalized numerical variables.
}
\description{
Normalize/Standardize Numerical Variables
}
