% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{pretraitement_complet}
\alias{pretraitement_complet}
\title{Full Preprocessing Pipeline}
\usage{
pretraitement_complet(
  data,
  cible,
  imputation_num = "moyenne",
  imputation_cat = "mode"
)
}
\arguments{
\item{data}{A dataframe containing the dataset to preprocess.}

\item{cible}{The name of the target variable (as a character string).}

\item{imputation_num}{The imputation method for numerical variables.}

\item{imputation_cat}{The imputation method for categorical variables.}
}
\value{
A fully preprocessed dataframe.
}
\description{
Full Preprocessing Pipeline
}
