% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{split_train_test}
\alias{split_train_test}
\title{Split Data into Training and Testing Sets}
\usage{
split_train_test(data, cible, proportion = 0.8)
}
\arguments{
\item{data}{A dataframe containing the dataset to split.}

\item{cible}{The name of the target variable (as a character string).}

\item{proportion}{The proportion of data to allocate to the training set.
Default is 0.8 (80\% training, 20\% testing).}
}
\value{
A list containing the training and testing sets.
}
\description{
Split Data into Training and Testing Sets
}
